Use Boost only on big endian and non-iec-559-platforms
