+++ Apps/LabRecorder/CMakeLists.txt
@@ -18,11 +18,14 @@ if(NOT TARGET LSL::lsl)
 	include(LSLCMake)
 endif()
 
+option(LABRECORDER_XDFZ "use Boost.Iostreams for XDFZ support" Off)
+option(LABRECORDER_BOOST_TYPE_CONVERSIONS "Use boost for type conversions" Off)
+
 # GENERAL CONFIG #
 set(META_PROJECT_DESCRIPTION "Record LabStreamingLayer streams to XDF data file.")
 
 # THIRD PARTY LIBRARIES #
-find_package(Qt5 REQUIRED COMPONENTS Network Widgets)
+find_package(Qt5 REQUIRED COMPONENTS Widgets)
 
 # APPLICATION #
 
@@ -39,37 +42,53 @@ add_executable(${target}
     mainwindow.ui
     recording.h
 	recording.cpp
+	conversions.h
 )
 target_compile_features(${target} PRIVATE cxx_auto_type cxx_lambda_init_captures)
 
 find_package(Threads REQUIRED)
-find_package(Boost REQUIRED OPTIONAL_COMPONENTS iostreams)
-if(WIN32)
-	find_package(Boost OPTIONAL_COMPONENTS zlib QUIET)
-	if(TARGET Boost::zlib)
-		set(ZLIB_LIBRARIES Boost::zlib)
-	endif()
-else()
-	find_package(ZLIB QUIET)
-endif()
 
 target_link_libraries(${target}
     PRIVATE
     Qt5::Widgets
-    Qt5::Network
-	Boost::boost
     Threads::Threads
     LSL::lsl
 )
 
+set(CMAKE_CXX_STANDARD 14)
+set(CMAKE_CXX_STANDARD_REQUIRED On)
+
+# Test for floating point format and endianness
+try_run(IS_LITTLE_ENDIAN IS_IEC559
+	${CMAKE_CURRENT_BINARY_DIR}
+	"${CMAKE_CURRENT_SOURCE_DIR}/test_iec559_and_little_endian.cpp"
+	CMAKE_FLAGS "-DCMAKE_CXX_STANDARD=14" "-DCMAKE_CXX_STANDARD_REQUIRED=On"
+	COMPILE_OUTPUT_VARIABLE IEC559_COMPILE)
+message(STATUS "Little endian: ${IS_LITTLE_ENDIAN}")
+message(STATUS "IEC559: ${IS_IEC559}")
+if(NOT IS_IEC559)
+	message(WARNING "IEC559 test: ${IEC559_COMPILE}")
+endif()
+
+
+if(LABRECORDER_BOOST_TYPE_CONVERSIONS OR NOT IS_LITTLE_ENDIAN OR NOT IS_IEC559)
+	message(STATUS "Trying to use Boost for type conversions")
+	find_package(Boost REQUIRED)
+	target_link_libraries(${target} PRIVATE Boost::boost)
+	target_compile_definitions(${target} PRIVATE EXOTIC_ARCH_SUPPORT)
+endif()
+
 # Enable xdfz support if Boost::iostreams and Boost.zlib (Windows) or plain zlib (Unix) was found
-if(ZLIB_LIBRARIES AND TARGET Boost::iostreams)
+if(LABRECORDER_XDFZ)
+	find_package(Boost REQUIRED COMPONENTS iostreams)
+	if(WIN32)
+		find_package(Boost REQUIRED COMPONENTS zlib)
+		target_link_libraries(${target} PRIVATE Boost::iostreams Boost::zlib)
+	else()
+		find_package(ZLIB REQUIRED)
+		target_link_libraries(${target} PRIVATE Boost::iostreams ${ZLIB_LIBRARIES})
+	endif()
 	message(STATUS "Found zlib, enabling support for xdfz files")
-	target_link_libraries(${target}
-		PRIVATE
-		${ZLIB_LIBRARIES}
-		Boost::iostreams
-		)
 	target_compile_definitions(${target} PRIVATE XDFZ_SUPPORT=1)
 endif()
